% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create_templates}
\alias{create_templates}
\title{Generate standardized \code{starter.ss} and \code{forecast.ss} files
for SS3 model setups}
\usage{
create_templates(modelfolder = ".", forecast = TRUE, starter = TRUE,
  verbose = FALSE)
}
\arguments{
  \item{modelfolder}{Path to a folder containing SS3
  models. Inside each model setup folder, there should be a
  folder named \code{om} and \code{em}. For example,
  \code{ss3models/inst/models} or
  \code{system.file("models", package = "ss3models")}.}

  \item{forecast}{A logical if \code{forecast.ss} files
  should be standardized.}

  \item{starter}{A logical if \code{starter.ss} files
  should be standardized.}

  \item{verbose}{A logical for debugging.}
}
\description{
Generate standardized \code{starter.ss} and
\code{forecast.ss} files for SS3 model setups
}
\examples{
\dontrun{
m <- system.file("models", package = "ss3models")
create_templates(m, forecast = TRUE, starter = TRUE)
}
}
\author{
Kelli Faye Johnson
}
\seealso{
\code{\link{make_forecast}}, \code{\link{make_starter}}
}

