% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_parvalues.R
\name{get_parvalues}
\alias{get_parvalues}
\title{Get parameter values for SS3 model setups}
\usage{
get_parvalues(modelfolder = ".", write_csv = TRUE,
  outfile = "parlist.csv", ss_binary = "ss3_24o_opt")
}
\arguments{
  \item{modelfolder}{Path to a folder containing SS3
  models. Inside each model setup folder, there should be a
  folder named \code{om} and \code{em}. For example,
  \code{ss3models/inst/models} or
  \code{system.file("models", package = "ss3models")}.}

  \item{write_csv}{Should the parameter value data frame be
  written to CSV file named \code{outfile}}

  \item{outfile}{A character value specifying a file name
  for the output if \code{write_csv = TRUE}. Must end in
  \code{.csv}.}

  \item{ss_binary}{A character value specifying which
  version of SS3 to use.}
}
\value{
A CSV file named \code{outfile} if \code{write_csv = TRUE}.
The parameter data frame is returned invisibly regardless.
}
\description{
Get parameter values for SS3 model setups
}
\examples{
\dontrun{
m <- system.file("models", package = "ss3models")
p <- get_parvalues(m, write_csv = FALSE)
head(p)
}
}
\author{
Kelli Faye Johnson
}

