% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ss3model.R
\name{list_models}
\alias{list_models}
\title{List all available model setups in the ss3models package}
\usage{
list_models()
}
\description{
List all available model setups in the ss3models package
}
\examples{
list_models()
}
\author{
Sean Anderson
}
\seealso{
\code{\link{ss3model}}
}

