% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_forecast.R
\name{make_forecast}
\alias{make_forecast}
\title{Write a \code{forecast.ss} file for use in \pkg{ss3sim}.}
\usage{
make_forecast(outfile = "forecast.ss", dir = NULL, msy = 2,
  spr_target = 0.4, b_target = 0.4, bmark_years = rep(0, 6),
  bmark_relf = 1, forecast = 0, nforecast = 0, forecast_years = rep(0,
  4), fleet = 1, verbose = FALSE, write_file = TRUE)
}
\arguments{
  \item{outfile}{A character value specifying the name to
  save the forecast file as. Will not be written to the
  disk unless \code{write_file = TRUE}. The default is
  \code{"forecast.ss"}.}

  \item{dir}{A directory to save \code{outfile} in. The
  default is \code{NULL}, where the \code{outfile} will be
  saved in \code{getwd()}.}

  \item{msy}{Specifies whether or not to do a forecast and
  which \emph{F} to use for that forecast. Use \code{0} to
  have no forecast, \code{1} to set to F(SPR), \code{2} to
  calculate F(MSY), \code{3} to set to F(Btarget), and
  \code{4} to set to F(endyr). Default is \code{2}.}

  \item{spr_target}{Stock Synthesis searches for an
  \emph{F} multiplier that will produce this level of
  spawning stock biomass (or reproductive output) per
  recruit relative to an unfished value. Default is
  \code{0.4}.}

  \item{b_target}{Stock Synthesis searches for an \emph{F}
  multiplier that will produce this level of spawning stock
  biomass relative to an unfished value. This is not per
  recruit and takes into account the Spawner-Recruit
  relationship. Default is \code{0.4}.}

  \item{bmark_years}{A vector of \code{length} 6: 1)
  beginning year of biology, 2) ending year of biology, 3)
  beginning year of selectivity, 4) ending year of
  selectivity, 5) beginning year of relative F, and ending
  year of relative F. All entries are less than or equal to
  zero, to set years relative to the terminal year of the
  model. Default is \code{c(0, 0, 0, 0, 0, 0)}.}

  \item{bmark_relf}{A numeric value of \code{1} or
  \code{2}, where the former instructs Stock Synthesis to
  use the year range from \code{bmark_years} to calculate a
  relative fishing mortality level and the later uses the
  same years as specified for the forecast in
  \code{fcast_years}.}

  \item{forecast}{Specifies whether or not to do a forecast
  benchmarks and which to calculate. Use \code{0} to have
  no forecast, \code{1} to set to F(SPR), \code{2} to
  calculate F(MSY), \code{3} to set to F(Btarget), and
  \code{4} to set to F(endyr). Default is \code{0}. Inputs
  are the same as \code{msy}.}

  \item{nforecast}{A numeric value specifying the number of
  forecast years. The default number of years is \code{0}.}

  \item{forecast_years}{A vector length, where each value
  corresponds to years relative to the end year of the
  simulation that will be used in the forecasts. Values
  correspond to the beginning and end years for selectivity
  and the relative \emph{F}s that will be used in
  population forecasts.}

  \item{fleet}{A numeric value assigning the forecast to a
  specific fleet in the model. Default is fleet \code{1}.}

  \item{verbose}{A logical specifying whether or not to
  write information to the console while writing the
  forecast files. Useful for debugging. Default is
  \code{FALSE}.}

  \item{write_file}{A logical specifying whether or not to
  write the file to the disk or not. File will always be
  returned as a \code{invisible} independent of
  \code{write_file}. Default is to write to the disk.}
}
\description{
Write a \code{forecast.ss} file for use in \pkg{ss3sim}.
}
\author{
Kelli Johnson
}

