% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_readme.R
\name{make_readme}
\alias{make_readme}
\title{Make a single README file}
\usage{
make_readme(path = ".")
}
\arguments{
  \item{path}{Folder path to the base of the package
  folder.}
}
\description{
Run this from within a model setup folder.
}

