% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_readme.R
\name{make_readmes}
\alias{make_readmes}
\title{Make all README files}
\usage{
make_readmes(path = ".")
}
\arguments{
  \item{path}{Folder path to the base package folder.}
}
\description{
Run this from the base package folder.
}

