% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_starter.R
\name{make_starter}
\alias{make_starter}
\title{Write a \code{starter.ss} file for use in \pkg{ss3sim}.}
\usage{
make_starter(outfile = "starter.ss", dir = NULL, type = c("om", "em"),
  last_estimation_phase = 15, verbose = FALSE, write_file = TRUE)
}
\arguments{
  \item{outfile}{A character value specifying the name to
  save the forecast file as. Will not be written to the
  disk unless \code{write_file = TRUE}. The default is
  \code{"forecast.ss"}.}

  \item{dir}{A directory to save \code{outfile} in. The
  default is \code{NULL}, where the \code{outfile} will be
  saved in \code{getwd()}.}

  \item{type}{A character value of either \code{"om"} or
  \code{"em"}, specifying if the \code{starter.ss} file
  pertains to an operating or estimating model set up.}

  \item{last_estimation_phase}{A numeric value specifying
  the last estimation phase in an estimation model. Not
  used if \code{type = "om"}.}

  \item{verbose}{A logical specifying whether or not to
  write information to the console while writing the
  forecast files. Useful for debugging. Default is
  \code{FALSE}.}

  \item{write_file}{A logical specifying whether or not to
  write the file to the disk or not. File will always be
  returned as a \code{invisible} independent of
  \code{write_file}. Default is to write to the disk.}
}
\description{
Write a \code{starter.ss} file for use in \pkg{ss3sim}.
}
\author{
Kelli Johnson
}

