% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ss3model.R
\name{ss3model}
\alias{ss3model}
\title{Get the system file path to a SS3 model}
\usage{
ss3model(species, type = c("om", "em"))
}
\arguments{
  \item{species}{Which model setup to choose. See the
  available model setups with \code{\link{list_models}}.}

  \item{type}{One of \code{"om"} or \code{"em"} for
  operating model or estimating model.}
}
\description{
Get the system file path to a SS3 model
}
\examples{
ss3model("hake", "em")
ss3model("flatfish", "om")
}
\author{
Sean Anderson
}
\seealso{
\code{\link{list_models}}
}

