% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check SS3 models}
\usage{
check_model(model, opts = "-nohess", ss_mode = c("safe", "optimized"))
}
\arguments{
  \item{model}{Path to a model to check}

  \item{opts}{Options to pass to SS3 on the command line}

  \item{ss_mode}{One of \code{"safe"} or \code{"optimized"}
  for which SS3 binary to run. See the ss3sim vignette for
  instructions.}
}
\description{
Check SS3 models
}

